refactor: share get_verification_hashes utility
